Exception handling - delete

+++ src/main/java/dev/andreia/vendas/services/UserService.java
@@ -2,8 +2,10 @@ package dev.andreia.vendas.services;
 
 import dev.andreia.vendas.entities.User;
 import dev.andreia.vendas.repositories.UserRepository;
+import dev.andreia.vendas.services.exceptions.DatabaseException;
 import dev.andreia.vendas.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,7 +31,16 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        if (!userRepository.existsById(id)){
+            throw new ResourceNotFoundException(id);
+        }
+
+        try{
+            userRepository.deleteById(id);
+        } catch (DataIntegrityViolationException ex){
+            throw new DatabaseException(ex.getMessage());
+        }
+
     }
 
     public User update(Long id, User obj){
